// domain/TrackByLabelId/types/ImageExampleKey.enum.ts
enum ImageExampleKey {
  CASTING_AND_RELEASE = 'Casting & Release Roll Label',
  EUROPEAN_GALERIE_ROLL = 'European Galerie Roll Label',
  EUROPEAN_GRAPHICS_PALLET = 'European Graphics Pallet Label',
  EUROPEAN_GRAPHICS_REAM = 'European Graphics Ream Label',
  EUROPEAN_GRAPHICS_ROLL = 'European Graphics Roll Label',
  NORTH_AMERICAN_PALLET = 'North American Pallet Label',
  NORTH_AMERICAN_ROLL = 'North American Roll Label',
}

export default ImageExampleKey;